+++ libvmi/os/windows/windows.h
@@ -76,7 +76,7 @@ windows_rva_to_export(vmi_instance_t vmi, addr_t rva, const access_context_t *ct
 status_t windows_teardown(vmi_instance_t vmi);
 
 typedef int (*check_magic_func)(uint32_t);
-int find_pname_offset(vmi_instance_t vmi, check_magic_func check);
+uint64_t find_pname_offset(vmi_instance_t vmi, check_magic_func check);
 addr_t windows_find_eprocess(vmi_instance_t instance, const char *name);
 addr_t eprocess_list_search(vmi_instance_t vmi, addr_t list_head, int offset, size_t len, void *value);
 addr_t windows_find_eprocess_list_pid(vmi_instance_t vmi, vmi_pid_t pid);
